move .md preprocessor into src file
